services/xml-service/requirements.txt
flask>=3.0.1
flask-cors>=4.0.0
psycopg2-binary>=2.9.9
lxml>=5.1.0
xmlschema>=2.5.1
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.1
python-dotenv>=1.0.0
requests>=2.31.0
